+++ node/typings/document.d.ts
@@ -1,7 +1,6 @@
 import { ID } from './scalars'
 
 export interface CacheableDocument {
-  /** id is used as cacheId */
   cacheId: ID
   id: ID
 }
@@ -16,18 +15,14 @@ export interface Field {
 }
 
 export interface DocumentPOSTResponse extends CacheableDocument {
-  /** documentId is used as cacheId */
   href: string
   documentId: string
 }
 
-/** Documents Input object to be used in Mutation */
-/** @graphql input */
 export interface DocumentInput {
   fields: FieldInput[]
 }
 
-/** @graphql input */
 export interface FieldInput {
   key: string
   value: string