+++ node/typings/scalars.d.ts
@@ -1,8 +1,5 @@
-/** @graphql ID */
 export type ID = string
 
-/** @graphql Int */
 export type Int = number
 
-/** @graphql Float */
 export type Float = number