+++ node/typings/schema-response.d.ts
@@ -32,19 +32,13 @@ export interface Properties {
   rcLastSessionDate: Property
 }
 
-/** CRM Schema*/
 export interface SchemaResponse {
-  /** Properties of Schema */
   properties: Properties
-  /** Required fields of Schema */
   required: string[]
-  /** Indexed fields of Schema */
   vIndexed: string[]
-  /** Custom fields created by the user */
   customFields?: CustomField[]
 }
 
-/** Custom field in a key/value format */
 export interface CustomField {
   key: string
   value: Property