+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.2] - 2022-08-10
+
 ## [3.0.1] - 2021-08-02
 ### Added
 - Readme with API documentation